feat(sipag-auth): extract identity / auth subsystem into its own crate
Step 8 in docs/extraction-plan.md §4 sequence (final scheduled
extraction). Mechanical move of sipag-core/src/auth/ (9 files:
mod.rs + credential / error / random / session / setup_token /
state / store / webauthn) into sipag-auth. `git mv` preserves
history.

What landed
-----------

- New `sipag-auth` workspace crate. Deps: anyhow, serde + json,
  thiserror, tempfile, tokio (sync+fs+macros), tracing, scrypt,
  rand_core, subtle, sha2, base64, url, uuid, webauthn-rs (with
  the `danger-allow-state-serialisation` feature — see in-crate
  doc for the safety story).
- All 9 files moved via `git mv` — board/mod.rs → src/lib.rs;
  the rest → src/<name>.rs.
- Bulk fix: `crate::auth::` → `crate::` (was 20+ internal
  references since `crate::` is now sipag-auth itself, not
  sipag-core).
- `sipag-core` re-exports at the old path via
  `pub use sipag_auth as auth`.
- `sipag` binary's `serve/error.rs` migrated to direct
  `use sipag_auth::AuthError` (satisfies §7 #3). Other callers
  in sipag/serve continue working through the shim.

Tests
-----

All 48 sipag-auth tests pass. `make dev` clean across the
workspace. This includes the heavier passkey/scrypt/setup-token
test suite (~3-4 minutes for the full pass — note the long-running
scrypt iterations).

[architecture] [extraction]
