+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "katulong-client",
     "ollama-bridge-client",
+    "sipag-auth",
     "sipag-board",
     "sipag-core",
     "sipag-corpus",
