+++ sipag-auth/src/error.rs
