+++ sipag/Cargo.toml
@@ -33,6 +33,10 @@ sipag-mesh = { path = "../sipag-mesh" }
 # re-exports at the old `sipag_core::board::…` path; new code
 # imports `sipag_board::…` directly.
 sipag-board = { path = "../sipag-board" }
+# Identity / auth subsystem. sipag-core re-exports at the old
+# `sipag_core::auth::…` path; new code imports `sipag_auth::…`
+# directly.
+sipag-auth = { path = "../sipag-auth" }
 clap = { version = "4", features = ["derive"] }
 anyhow = "1"
 serde = { version = "1", features = ["derive"] }
